Adding autocmd description and fixing bug on last commit
After the first time leaving the Fugitive buffer, since which-key now
had a "which_key_ignore" tag associated with that key combination, it
stopped appearing on which-key even when it should. The solution
provided might not be ideal, but it is good enough while which-key
doesn't add functionality to remove a keybind.